requirements.txt
Click>=8.0
PyYAML>=5.4
typer>=0.4.1
psutil>=5.8
asyncio>=3.4

